+++ Remaining_Shell/Win32Explorer_26.0.3.0/App_Source/ToolbarHelper.h
@@ -1,4 +1,4 @@
-﻿// Copyright (C) Win32Explorer Project
+// Copyright (C) Win32Explorer Project
 // SPDX-License-Identifier: GPL-3.0-only
 // See LICENSE in the top level directory
 
@@ -18,7 +18,6 @@ namespace ToolbarHelper
 static constexpr int CLOSE_TOOLBAR_X_OFFSET = 4;
 
 std::tuple<HWND, std::vector<wil::unique_himagelist>> CreateCloseButtonToolbar(HWND parent,
-	int closeButtonId, const std::wstring &tooltip, const ResourceLoader *resourceLoader);
+	int buttonId, const std::wstring &tooltipText, const ResourceLoader *resourceLoader, int iconSize = 24, bool isTab = false);
 
 }
-